requirements.txt
azure-iot-device==2.12.0
azure-iot-hub==2.6.1
azure-eventhub==5.11.4
azure-functions==1.18.0
azure-storage-blob==12.19.0
redis==5.0.1
pyodbc==5.0.1
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
paho-mqtt==1.6.1
